split ClearButton and Colors into separate files
